+++ packages/design-system/src/components/ds-table/filters/hooks/index.ts
@@ -1,2 +1,2 @@
 export { useTableFilters } from './use-table-filters';
-export type { UseTableFiltersResult } from './use-table-filters';
+export type { UseTableFiltersOptions, UseTableFiltersResult } from './use-table-filters';